Replace CollectionView with NextCollectionView
Getting the ball rolling.
This also removes CompositeView.
It can be re-added in another PR.
