+++ src/collection-view.js
@@ -7,6 +7,7 @@ import { renderView, destroyView } from './common/view';
 import monitorViewEvents from './common/monitor-view-events';
 import ChildViewContainer from './child-view-container';
 import MarionetteError from './error';
+import Region from './region';
 import ViewMixin from './mixins/view';
 import { setDomApi } from './config/dom';
 
@@ -18,32 +19,24 @@ const ClassOptions = [
   'childViewOptions',
   'childViewTriggers',
   'collectionEvents',
-  'events',
-  'filter',
   'emptyView',
   'emptyViewOptions',
+  'events',
   'modelEvents',
-  'reorderOnSort',
-  'sort',
+  'sortWithCollection',
   'triggers',
   'ui',
-  'viewComparator'
+  'viewComparator',
+  'viewFilter'
 ];
 
 // A view that iterates over a Backbone.Collection
 // and renders an individual child view for each model.
 const CollectionView = Backbone.View.extend({
-
   // flag for maintaining the sorted order of the collection
-  sort: true,
+  sortWithCollection: true,
 
   // constructor
-  // option to pass `{sort: false}` to prevent the `CollectionView` from
-  // maintaining the sorted order of the collection.
-  // This will fallback onto appending childView's to the end.
-  //
-  // option to pass `{viewComparator: compFunction()}` to allow the `CollectionView`
-  // to use a custom sort order for the collection.
   constructor(options) {
     this._setOptions(options);
 
@@ -51,438 +44,144 @@ const CollectionView = Backbone.View.extend({
 
     monitorViewEvents(this);
 
-    this._initBehaviors();
     this.once('render', this._initialEvents);
+
+    // This children container isn't really used by a render, but it provides
+    // the ability to check `this.children.length` prior to rendering
+    // It also allows for cases where only addChildView is used
     this._initChildViewStorage();
-    this._bufferedChildren = [];
+    this._initBehaviors();
 
     const args = Array.prototype.slice.call(arguments);
     args[0] = this.options;
     Backbone.View.prototype.constructor.apply(this, args);
+
+    // Init empty region
+    this.getEmptyRegion();
 
     this.delegateEntityEvents();
 
     this._triggerEventOnBehaviors('initialize', this);
   },
 
-  // Instead of inserting elements one by one into the page, it's much more performant to insert
-  // elements into a document fragment and then insert that document fragment into the page
-  _startBuffering() {
-    this._isBuffering = true;
+  // Internal method to set up the `children` object for storing all of the child views
+  _initChildViewStorage() {
+    this.children = new ChildViewContainer();
   },
 
-  _endBuffering() {
-    const shouldTriggerAttach = this._isAttached && this.monitorViewEvents !== false;
-    const triggerOnChildren = shouldTriggerAttach ? this._getImmediateChildren() : [];
+  // Create an region to show the emptyView
+  getEmptyRegion() {
+    if (this._emptyRegion && !this._emptyRegion.isDestroyed()) {
+      return this._emptyRegion;
+    }
 
-    this._isBuffering = false;
+    this._emptyRegion = new Region({ el: this.el, replaceElement: false });
 
-    _.each(triggerOnChildren, child => {
-      child.triggerMethod('before:attach', child);
-    });
+    this._emptyRegion._parentView = this;
 
-    this.attachBuffer(this, this._createBuffer());
-
-    _.each(triggerOnChildren, child => {
-      child._isAttached = true;
-      child.triggerMethod('attach', child);
-    });
-
-    this._bufferedChildren = [];
-  },
-
-  _getImmediateChildren() {
-    return _.values(this.children._views);
+    return this._emptyRegion;
   },
 
   // Configured the initial events that the collection view binds to.
   _initialEvents() {
-    if (this.collection) {
-      this.listenTo(this.collection, 'add', this._onCollectionAdd);
-      this.listenTo(this.collection, 'update', this._onCollectionUpdate);
-      this.listenTo(this.collection, 'reset', this.render);
-
-      if (this.sort) {
-        this.listenTo(this.collection, 'sort', this._sortViews);
-      }
-    }
-  },
-
-  // Handle a child added to the collection
-  _onCollectionAdd(child, collection, opts) {
-    // `index` is present when adding with `at` since BB 1.2; indexOf fallback for < 1.2
-    let index = opts.at !== undefined && (opts.index || collection.indexOf(child));
-
-    // When filtered or when there is no initial index, calculate index.
-    if (this.filter || index === false) {
-      index = _.indexOf(this._filteredSortedModels(index), child);
-    }
-
-    if (this._shouldAddChild(child, index)) {
-      this._destroyEmptyView();
-      this._addChild(child, index)
-    }
-  },
-
-  // Handle collection update model removals
-  _onCollectionUpdate(collection, options) {
-    const changes = options.changes;
-    this._removeChildModels(changes.removed);
-  },
-
-  // Remove the child views and destroy them.
-  // This function also updates the indices of later views
-  // in the collection in order to keep the children in sync with the collection.
-  // "models" is an array of models and the corresponding views
-  // will be removed and destroyed from the CollectionView
-  _removeChildModels(models) {
-    // Used to determine where to update the remaining
-    // sibling view indices after these views are removed.
-    const removedViews = this._getRemovedViews(models);
-
-    if (!removedViews.length) {
-      return;
-    }
-
-    this.children._updateLength();
-
-    // decrement the index of views after this one
-    this._updateIndices(removedViews, false);
-
-    if (this.isEmpty()) {
-      this._showEmptyView();
-    }
-  },
-
-  // Returns the views that will be used for re-indexing
-  // through CollectionView#_updateIndices.
-  _getRemovedViews(models) {
-
-    // Returning a view means something was removed.
-    return _.reduce(models, (removingViews, model) => {
-      const view = model && this.children.findByModel(model);
-
-      if (!view || view._isDestroyed) {
-        return removingViews;
-      }
-
-      this._removeChildView(view);
-
-      removingViews.push(view);
-
-      return removingViews;
-    }, []);
-  },
-
-  _removeChildView(view) {
-    this.triggerMethod('before:remove:child', this, view);
-
-    this.children._remove(view);
-    view._shouldDisableEvents = this.monitorViewEvents === false;
-    destroyView(view);
-
-    this.stopListening(view);
-    this.triggerMethod('remove:child', this, view);
-  },
-
-  // Overriding Backbone.View's `setElement` to handle
-  // if an el was previously defined. If so, the view might be
-  // attached on setElement.
-  setElement() {
-    Backbone.View.prototype.setElement.apply(this, arguments);
-
-    this._isAttached = this.Dom.hasEl(document.documentElement, this.el);
-
-    return this;
-  },
-
-  // Render children views. Override this method to provide your own implementation of a
-  // render function for the collection view.
-  render() {
-    if (this._isDestroyed) { return this; }
-    this.triggerMethod('before:render', this);
-    this._renderChildren();
-    this._isRendered = true;
-    this.triggerMethod('render', this);
-    return this;
-  },
-
-  // An efficient rendering used for filtering. Instead of modifying the whole DOM for the
-  // collection view, we are only adding or removing the related childrenViews.
-  setFilter(filter, {preventRender} = {}) {
-    const canBeRendered = this._isRendered && !this._isDestroyed;
-    const filterChanged = this.filter !== filter;
-    const shouldRender = canBeRendered && filterChanged && !preventRender;
-
-    if (shouldRender) {
-      const previousModels = this._filteredSortedModels();
-      this.filter = filter;
-      const models = this._filteredSortedModels();
-      this._applyModelDeltas(models, previousModels);
-    } else {
-      this.filter = filter;
-    }
-
-    return this;
-  },
-
-  // `removeFilter` is actually an alias for removing filters.
-  removeFilter(options) {
-    return this.setFilter(null, options);
-  },
-
-  // Calculate and apply difference by cid between `models` and `previousModels`.
-  _applyModelDeltas(models, previousModels) {
-    const currentIds = {};
-    _.each(models, (model, index) => {
-      const addedChildNotExists = !this.children.findByModel(model);
-      if (addedChildNotExists) {
-        this._onCollectionAdd(model, this.collection, {at: index});
-      }
-      currentIds[model.cid] = true;
+    this.listenTo(this.collection, {
+      'sort': this._onCollectionSort,
+      'reset': this._onCollectionReset,
+      'update': this._onCollectionUpdate
     });
-
-    const removeModels = _.filter(previousModels, (prevModel) => {
-      return !currentIds[prevModel.cid] && this.children.findByModel(prevModel);
-    });
-
-    this._removeChildModels(removeModels);
-  },
-
-  // Reorder DOM after sorting. When your element's rendering do not use their index,
-  // you can pass reorderOnSort: true to only reorder the DOM after a sort instead of
-  // rendering all the collectionView.
-  reorder() {
-    const children = this.children;
-    const models = this._filteredSortedModels();
-
-    if (!models.length && this._showingEmptyView) { return this; }
-
-    const anyModelsAdded = _.some(models, function(model) {
-      return !children.findByModel(model);
-    });
-
-    // If there are any new models added due to filtering we need to add child views,
-    // so render as normal.
-    if (anyModelsAdded) {
-      this.render();
-    } else {
-
-      const filteredOutModels = [];
-
-      // Get the DOM nodes in the same order as the models and
-      // find the model that were children before but aren't in this new order.
-      const elsToReorder = _.reduce(this.children._views, function(viewEls, view) {
-        const index = _.indexOf(models, view.model);
-
-        if (index === -1) {
-          filteredOutModels.push(view.model);
-          return viewEls;
-        }
-
-        view._index = index;
-
-        viewEls[index] = view.el;
-
-        return viewEls;
-      }, new Array(models.length));
-
-      this.triggerMethod('before:reorder', this);
-
-      const elBuffer = this.Dom.createBuffer();
-
-      _.each(elsToReorder, (el) => {
-        this.Dom.appendContents(elBuffer, el);
-      });
-
-      // Since append moves elements that are already in the DOM, appending the elements
-      // will effectively reorder them.
-      this._appendReorderedChildren(elBuffer);
-
-      // remove any views that have been filtered out
-      this._removeChildModels(filteredOutModels);
-
-      this.triggerMethod('reorder', this);
-    }
-    return this;
-  },
-
-  // Render view after sorting. Override this method to change how the view renders
-  // after a `sort` on the collection.
-  resortView() {
-    if (this.reorderOnSort) {
-      this.reorder();
-    } else {
-      this._renderChildren();
-    }
-    return this;
   },
 
   // Internal method. This checks for any changes in the order of the collection.
-  // If the index of any view doesn't match, it will render.
-  _sortViews() {
-    const models = this._filteredSortedModels();
-
-    // check for any changes in sort order of views
-    const orderChanged = _.find(models, (item, index) => {
-      const view = this.children.findByModel(item);
-      return !view || view._index !== index;
-    });
-
-    if (orderChanged) {
-      this.resortView();
-    }
-  },
-
-  // Internal reference to what index a `emptyView` is.
-  _emptyViewIndex: -1,
-
-  // Internal method. Separated so that CompositeView can append to the childViewContainer
-  // if necessary
-  _appendReorderedChildren(children) {
-    this.Dom.appendContents(this.el, children, {_$el: this.$el});
-  },
-
-  // Internal method. Separated so that CompositeView can have more control over events
-  // being triggered, around the rendering process
-  _renderChildren() {
-    if (this._isRendered) {
-      this._destroyEmptyView();
-      this._destroyChildren();
+  // If the index of any view doesn't match, it will re-sort.
+  _onCollectionSort(collection, { add, merge, remove }) {
+    if (!this.sortWithCollection || this.viewComparator === false) {
+      return;
     }
 
-    const models = this._filteredSortedModels();
-    if (this.isEmpty({processedModels: models})) {
-      this._showEmptyView();
-    } else {
-      this.triggerMethod('before:render:children', this);
-      this._startBuffering();
-      this._showCollection(models);
-      this._endBuffering();
-      this.triggerMethod('render:children', this);
+    // If the data is changing we will handle the sort later in `_onCollectionUpdate`
+    if (add || remove || merge) {
+      return;
     }
+
+    // If the only thing happening here is sorting, sort.
+    this.sort();
   },
 
-  _createView(model, index) {
-    const ChildView = this._getChildView(model);
-    const childViewOptions = this._getChildViewOptions(model, index);
-    const view = this.buildChildView(model, ChildView, childViewOptions);
+  _onCollectionReset() {
+    this.render();
+  },
+
+  // Handle collection update model additions and  removals
+  _onCollectionUpdate(collection, options) {
+    const changes = options.changes;
+
+    // Remove first since it'll be a shorter array lookup.
+    const removedViews = changes.removed.length && this._removeChildModels(changes.removed);
+
+    this._addedViews = changes.added.length && this._addChildModels(changes.added);
+
+    this._detachChildren(removedViews);
+
+    this._showChildren();
+
+    // Destroy removed child views after all of the render is complete
+    this._removeChildViews(removedViews);
+  },
+
+  _removeChildModels(models) {
+    return _.reduce(models, (views, model) => {
+      const removeView = this._removeChildModel(model);
+
+      if (removeView) { views.push(removeView); }
+
+      return views;
+    }, []);
+  },
+
+  _removeChildModel(model) {
+    const view = this.children.findByModel(model);
+
+    if (view) { this._removeChild(view); }
+
     return view;
   },
 
-  _setupChildView(view, index) {
-    monitorViewEvents(view);
+  _removeChild(view) {
+    this.triggerMethod('before:remove:child', this, view);
 
-    // set up the child view event forwarding
-    this._proxyChildViewEvents(view);
+    this.children._remove(view);
 
-    if (this.sort) {
-      view._index = index;
-    }
+    this.triggerMethod('remove:child', this, view);
   },
 
-  // Internal method to loop through collection and show each child view.
-  _showCollection(models) {
-    _.each(models, _.bind(this._addChild, this));
-    this.children._updateLength();
+  // Added views are returned for consistency with _removeChildModels
+  _addChildModels(models) {
+    return _.map(models, _.bind(this._addChildModel, this));
   },
 
-  // Allow the collection to be sorted by a custom view comparator
-  _filteredSortedModels(addedAt) {
-    if (!this.collection || !this.collection.length) { return []; }
+  _addChildModel(model) {
+    const view = this._createChildView(model);
 
-    const viewComparator = this.getViewComparator();
-    let models = this.collection.models;
-    addedAt = Math.min(Math.max(addedAt, 0), models.length - 1);
+    this._addChild(view);
 
-    if (viewComparator) {
-      let addedModel;
-      // Preserve `at` location, even for a sorted view
-      if (addedAt) {
-        addedModel = models[addedAt];
-        models = models.slice(0, addedAt).concat(models.slice(addedAt + 1));
-      }
-      models = this._sortModelsBy(models, viewComparator);
-      if (addedModel) {
-        models.splice(addedAt, 0, addedModel);
-      }
-    }
-
-    // Filter after sorting in case the filter uses the index
-    models = this._filterModels(models);
-
-    return models;
+    return view;
   },
 
-  getViewComparator() {
-    return this.viewComparator;
+  _createChildView(model) {
+    const ChildView = this._getChildView(model);
+    const childViewOptions = this._getChildViewOptions(model);
+    const view = this.buildChildView(model, ChildView, childViewOptions);
+
+    return view;
   },
 
-  // Filter an array of models, if a filter exists
-  _filterModels(models) {
-    if (this.filter) {
-      models = _.filter(models, (model, index) => {
-        return this._shouldAddChild(model, index);
-      });
-    }
-    return models;
-  },
+  _addChild(view, index) {
+    this.triggerMethod('before:add:child', this, view);
 
-  _sortModelsBy(models, comparator) {
-    if (typeof comparator === 'string') {
-      return _.sortBy(models, (model) => {
-        return model.get(comparator);
-      });
-    } else if (comparator.length === 1) {
-      return _.sortBy(models, _.bind(comparator, this));
-    } else {
-      return _.clone(models).sort(_.bind(comparator, this));
-    }
-  },
+    this._setupChildView(view);
+    this.children._add(view, index);
 
-  // Internal method to show an empty view in place of a collection of child views,
-  // when the collection is empty
-  _showEmptyView() {
-    const EmptyView = this._getEmptyView();
-
-    if (EmptyView && !this._showingEmptyView) {
-      this._showingEmptyView = true;
-
-      const model = new Backbone.Model();
-      let emptyViewOptions =
-        this.emptyViewOptions || this.childViewOptions;
-      if (_.isFunction(emptyViewOptions)) {
-        emptyViewOptions = emptyViewOptions.call(this, model, this._emptyViewIndex);
-      }
-
-      const view = this.buildChildView(model, EmptyView, emptyViewOptions);
-
-      this.triggerMethod('before:render:empty', this, view);
-      this.addChildView(view, 0);
-      this.triggerMethod('render:empty', this, view);
-    }
-  },
-
-  // Internal method to destroy an existing emptyView instance if one exists. Called when
-  // a collection view has been rendered empty, and then a child is added to the collection.
-  _destroyEmptyView() {
-    if (this._showingEmptyView) {
-      this.triggerMethod('before:remove:empty', this);
-
-      this._destroyChildren();
-      delete this._showingEmptyView;
-
-      this.triggerMethod('remove:empty', this);
-    }
-  },
-
-  // Retrieve the empty view class
-  _getEmptyView() {
-    const emptyView = this.emptyView;
-
-    if (!emptyView) { return; }
-
-    return this._getView(emptyView);
+    this.triggerMethod('add:child', this, view);
   },
 
   // Retrieve the `childView` class
@@ -521,215 +220,484 @@ const CollectionView = Backbone.View.extend({
     }
   },
 
-  // Internal method for building and adding a child view
-  _addChild(child, index) {
-    const view = this._createView(child, index);
-    this.addChildView(view, index);
-
-    return view;
-  },
-
-  _getChildViewOptions(child, index) {
+  _getChildViewOptions(child) {
     if (_.isFunction(this.childViewOptions)) {
-      return this.childViewOptions(child, index);
+      return this.childViewOptions(child);
     }
 
     return this.childViewOptions;
   },
 
-  // Render the child's view and add it to the HTML for the collection view at a given index.
-  // This will also update the indices of later views in the collection in order to keep the
-  // children in sync with the collection.
-  addChildView(view, index) {
-    this.triggerMethod('before:add:child', this, view);
-    this._setupChildView(view, index);
-
-    // Store the child view itself so we can properly remove and/or destroy it later
-    if (this._isBuffering) {
-      // Add to children, but don't update children's length.
-      this.children._add(view);
-    } else {
-      // increment indices of views after this one
-      this._updateIndices(view, true);
-      this.children.add(view);
-    }
-
-    renderView(view);
-
-    this._attachView(view, index);
-
-    this.triggerMethod('add:child', this, view);
-
-    return view;
-  },
-
-  // Internal method. This decrements or increments the indices of views after the added/removed
-  // view to keep in sync with the collection.
-  _updateIndices(views, increment) {
-    if (!this.sort) {
-      return;
-    }
-
-    if (!increment) {
-      _.each(_.sortBy(this.children._views, '_index'), (view, index) => {
-        view._index = index;
-      });
-
-      return;
-    }
-
-    const view = _.isArray(views) ? _.max(views, '_index') : views;
-
-    if (_.isObject(view)) {
-      // update the indexes of views after this one
-      _.each(this.children._views, (laterView) => {
-        if (laterView._index >= view._index) {
-          laterView._index += 1;
-        }
-      });
-    }
-  },
-
-  _attachView(view, index) {
-    // Only trigger attach if already attached and not buffering,
-    // otherwise _endBuffering() or Region#show() handles this.
-    const shouldTriggerAttach = !view._isAttached && !this._isBuffering && this._isAttached && this.monitorViewEvents !== false;
-
-    if (shouldTriggerAttach) {
-      view.triggerMethod('before:attach', view);
-    }
-
-    this.attachHtml(this, view, index);
-
-    if (shouldTriggerAttach) {
-      view._isAttached = true;
-      view.triggerMethod('attach', view);
-    }
-  },
-
   // Build a `childView` for a model in the collection.
+  // Override to customize the build
   buildChildView(child, ChildViewClass, childViewOptions) {
     const options = _.extend({model: child}, childViewOptions);
     return new ChildViewClass(options);
   },
 
-  // Remove the child view and destroy it. This function also updates the indices of later views
-  // in the collection in order to keep the children in sync with the collection.
-  removeChildView(view) {
+  _setupChildView(view) {
+    monitorViewEvents(view);
+
+    // We need to listen for if a view is destroyed in a way other
+    // than through the CollectionView.
+    // If this happens we need to remove the reference to the view
+    // since once a view has been destroyed we can not reuse it.
+    view.on('destroy', this.removeChildView, this);
+
+    // set up the child view event forwarding
+    this._proxyChildViewEvents(view);
+  },
+
+  // used by ViewMixin's `_childViewEventHandler`
+  _getImmediateChildren() {
+    return this.children._views;
+  },
+
+  // Overriding Backbone.View's `setElement` to handle
+  // if an el was previously defined. If so, the view might be
+  // attached on setElement.
+  setElement() {
+    Backbone.View.prototype.setElement.apply(this, arguments);
+
+    this._isAttached = this.Dom.hasEl(document.documentElement, this.el);
+
+    return this;
+  },
+
+  // Render children views.
+  render() {
+    if (this._isDestroyed) { return this; }
+    this.triggerMethod('before:render', this);
+
+    this._destroyChildren();
+
+    // After all children have been destroyed re-init the container
+    this.children._init();
+
+    if (this.collection) {
+      this._addChildModels(this.collection.models);
+    }
+
+    this._showChildren();
+
+    this._isRendered = true;
+
+    this.triggerMethod('render', this);
+    return this;
+  },
+
+  // Sorts the children then filters and renders the results.
+  sort() {
+    if (this._isDestroyed) { return this; }
+
+    if (!this.children.length) { return this; }
+
+    this._showChildren();
+
+    return this;
+  },
+
+  _showChildren() {
+    if (this.isEmpty()) {
+      this._showEmptyView();
+      return;
+    }
+
+    this._sortChildren();
+
+    this.filter();
+  },
+
+  // Returns true if the collectionView is considered empty.
+  // This is called twice during a render. Once to check the data,
+  // and again when views are filtered. Override this function to
+  // customize what empty means.
+  isEmpty(allViewsFiltered) {
+    return allViewsFiltered || !this.children.length;
+  },
+
+  _showEmptyView() {
+    const EmptyView = this._getEmptyView();
+
+    if (!EmptyView) {
+      return;
+    }
+
+    const options = this._getEmptyViewOptions();
+
+    const emptyRegion = this.getEmptyRegion();
+
+    emptyRegion.show(new EmptyView(options));
+  },
+
+  // Retrieve the empty view class
+  _getEmptyView() {
+    const emptyView = this.emptyView;
+
+    if (!emptyView) { return; }
+
+    return this._getView(emptyView);
+  },
+
+  // Remove the emptyView
+  _destroyEmptyView() {
+    const emptyRegion = this.getEmptyRegion();
+    // Only empty if a view is show so the region
+    // doesn't detach any other unrelated HTML
+    if (emptyRegion.hasView()) {
+      emptyRegion.empty();
+    }
+  },
+
+  //
+  _getEmptyViewOptions() {
+    const emptyViewOptions = this.emptyViewOptions || this.childViewOptions;
+
+    if (_.isFunction(emptyViewOptions)) {
+      return emptyViewOptions.call(this);
+    }
+
+    return emptyViewOptions;
+  },
+
+  // Sorts views by viewComparator and sets the children to the new order
+  _sortChildren() {
+    let viewComparator = this.getComparator();
+
+    if (!viewComparator) { return; }
+
+    this.triggerMethod('before:sort', this);
+
+    if (_.isFunction(viewComparator)) {
+      // Must use native bind to preserve length
+      viewComparator = viewComparator.bind(this);
+    }
+
+    this.children._sort(viewComparator);
+
+    this.triggerMethod('sort', this);
+  },
+
+  // Sets the view's `viewComparator` and applies the sort if the view is ready.
+  // To prevent the render pass `{ preventRender: true }` as the 2nd argument.
+  setComparator(comparator, {preventRender} = {}) {
+    const comparatorChanged = this.viewComparator !== comparator;
+    const shouldSort = comparatorChanged && !preventRender;
+
+    this.viewComparator = comparator;
+
+    if (shouldSort) {
+      this.sort();
+    }
+
+    return this;
+  },
+
+  // Clears the `viewComparator` and follows the same rules for rendering as `setComparator`.
+  removeComparator(options) {
+    return this.setComparator(null, options);
+  },
+
+  // If viewComparator is overriden it will be returned here.
+  // Additionally override this function to provide custom
+  // viewComparator logic
+  getComparator() {
+    if (this.viewComparator) { return this.viewComparator }
+
+    if (!this.sortWithCollection || this.viewComparator === false || !this.collection) {
+      return false;
+    }
+
+    return this._viewComparator;
+  },
+
+  // Default internal view comparator that order the views by
+  // the order of the collection
+  _viewComparator(view) {
+    return this.collection.indexOf(view.model);
+  },
+
+  // This method re-filters the children views and re-renders the results
+  filter() {
+    if (this._isDestroyed) { return this; }
+
+    if (!this.children.length) { return this; }
+
+    const filteredViews = this._filterChildren();
+
+    this._renderChildren(filteredViews);
+
+    return this;
+  },
+
+  _isAddedAtEnd(addedView, index, addedViews) {
+    const viewIndex = this.children._views.length - addedViews.length + index;
+    return addedView === this.children._views[viewIndex];
+  },
+
+  _filterChildren() {
+    const viewFilter = this._getFilter();
+    const addedViews = this._addedViews;
+
+    delete this._addedViews;
+
+    if (!viewFilter) {
+      if (!this.sortWithCollection && addedViews && _.every(addedViews, _.bind(this._isAddedAtEnd, this))) {
+        return addedViews;
+      }
+
+      return this.children._views;
+    }
+
+    this.triggerMethod('before:filter', this);
+
+    const filteredViews = _.partition(this.children._views, _.bind(viewFilter, this));
+
+    this._detachChildren(filteredViews[1]);
+
+    this.triggerMethod('filter', this);
+
+    return filteredViews[0];
+  },
+
+  // This method returns a function for the viewFilter
+  _getFilter() {
+    const viewFilter = this.getFilter();
+
+    if (!viewFilter) { return false; }
+
+    if (_.isFunction(viewFilter)) {
+      return viewFilter;
+    }
+
+    // Support filter predicates `{ fooFlag: true }`
+    if (_.isObject(viewFilter)) {
+      const matcher = _.matches(viewFilter);
+      return function(view) {
+        return matcher(view.model && view.model.attributes);
+      };
+    }
+
+    // Filter by model attribute
+    if (_.isString(viewFilter)) {
+      return function(view) {
+        return view.model && view.model.get(viewFilter);
+      };
+    }
+
+    throw new MarionetteError({
+      name: 'InvalidViewFilterError',
+      message: '"viewFilter" must be a function, predicate object literal, a string indicating a model attribute, or falsy'
+    });
+  },
+
+  // Override this function to provide custom
+  // viewFilter logic
+  getFilter() {
+    return this.viewFilter;
+  },
+
+  // Sets the view's `viewFilter` and applies the filter if the view is ready.
+  // To prevent the render pass `{ preventRender: true }` as the 2nd argument.
+  setFilter(filter, {preventRender} = {}) {
+    const filterChanged = this.viewFilter !== filter;
+    const shouldRender = filterChanged && !preventRender;
+
+    this.viewFilter = filter;
+
+    if (shouldRender) {
+      this.filter();
+    }
+
+    return this;
+  },
+
+  // Clears the `viewFilter` and follows the same rules for rendering as `setFilter`.
+  removeFilter(options) {
+    return this.setFilter(null, options);
+  },
+
+  _detachChildren(detachingViews) {
+    _.each(detachingViews, _.bind(this._detachChildView, this));
+  },
+
+  _detachChildView(view) {
+    const shouldTriggerDetach = view._isAttached && this.monitorViewEvents !== false;
+    if (shouldTriggerDetach) {
+      view.triggerMethod('before:detach', view);
+    }
+
+    this.detachHtml(view);
+
+    if (shouldTriggerDetach) {
+      view._isAttached = false;
+      view.triggerMethod('detach', view);
+    }
+  },
+
+  // Override this method to change how the collectionView detaches a child view
+  detachHtml(view) {
+    this.Dom.detachEl(view.el, view.$el);
+  },
+
+  _renderChildren(views) {
+    if (this.isEmpty(!views.length)) {
+      this._showEmptyView();
+      return;
+    }
+
+    this._destroyEmptyView();
+
+    this.triggerMethod('before:render:children', this, views);
+
+    const els = this._getBuffer(views);
+
+    this._attachChildren(els, views);
+
+    this.triggerMethod('render:children', this, views);
+  },
+
+  _attachChildren(els, views) {
+    const shouldTriggerAttach = this._isAttached && this.monitorViewEvents !== false;
+
+    views = shouldTriggerAttach ? views : [];
+
+    _.each(views, view => {
+      if (view._isAttached) { return; }
+      view.triggerMethod('before:attach', view);
+    });
+
+    this.attachHtml(els);
+
+    _.each(views, view => {
+      if (view._isAttached) { return; }
+      view._isAttached = true;
+      view.triggerMethod('attach', view);
+    });
+  },
+
+  // Renders each view in children and creates a fragment buffer from them
+  _getBuffer(views) {
+    const elBuffer = this.Dom.createBuffer();
+
+    _.each(views, view => {
+      renderView(view);
+      this.Dom.appendContents(elBuffer, view.el, {_$contents: view.$el});
+    });
+
+    return elBuffer;
+  },
+
+  // Override this method to do something other than `.append`.
+  // You can attach any HTML at this point including the els.
+  attachHtml(els) {
+    this.Dom.appendContents(this.el, els, {_$el: this.$el});
+  },
+
+  swapChildViews(view1, view2) {
+    if (!this.children.hasView(view1) || !this.children.hasView(view2)) {
+      throw new MarionetteError({
+        name: 'ChildSwapError',
+        message: 'Both views must be children of the collection view'
+      });
+    }
+
+    this.children._swap(view1, view2);
+    this.Dom.swapEl(view1.el, view2.el);
+
+    // If the views are not filtered the same, refilter
+    if (this.Dom.hasEl(this.el, view1.el) !== this.Dom.hasEl(this.el, view2.el)) {
+      this.filter();
+    }
+
+    return this;
+  },
+
+  // Render the child's view and add it to the HTML for the collection view at a given index, based on the current sort
+  addChildView(view, index) {
     if (!view || view._isDestroyed) {
       return view;
     }
 
-    this._removeChildView(view);
-    this.children._updateLength();
-    // decrement the index of views after this one
-    this._updateIndices(view, false);
+    this._addChild(view, index);
+    this._addedViews = [view];
+    this._showChildren();
+
     return view;
   },
 
-  // check if the collection is empty or optionally whether an array of pre-processed models is empty
-  isEmpty(options) {
-    let models;
-    if (_.result(options, 'processedModels')) {
-      models = options.processedModels;
+  // Detach a view from the children.  Best used when adding a
+  // childView from `addChildView`
+  detachChildView(view) {
+    this.removeChildView(view, { shouldDetach: true });
+
+    return view;
+  },
+
+  // Remove the child view and destroy it.  Best used when adding a
+  // childView from `addChildView`
+  // The options argument is for internal use only
+  removeChildView(view, options) {
+    if (!view) {
+      return view;
+    }
+
+    this._removeChildView(view, options);
+
+    this._removeChild(view);
+
+    if (this.isEmpty()) {
+      this._showEmptyView();
+    }
+
+    return view;
+  },
+
+  _removeChildViews(views) {
+    _.each(views, _.bind(this._removeChildView, this));
+  },
+
+  _removeChildView(view, {shouldDetach} = {}) {
+    view.off('destroy', this.removeChildView, this);
+
+    if (shouldDetach) {
+      this._detachChildView(view);
     } else {
-      models = this.collection ? this.collection.models : [];
-      models = this._filterModels(models);
-    }
-    return models.length === 0;
-  },
-
-  // You might need to override this if you've overridden attachHtml
-  attachBuffer(collectionView, buffer) {
-    this.Dom.appendContents(collectionView.el, buffer, {_$el: collectionView.$el});
-  },
-
-  // Create a fragment buffer from the currently buffered children
-  _createBuffer() {
-    const elBuffer = this.Dom.createBuffer();
-    _.each(this._bufferedChildren, (b) => {
-      this.Dom.appendContents(elBuffer, b.el, {_$contents: b.$el});
-    });
-    return elBuffer;
-  },
-
-  // Append the HTML to the collection's `el`. Override this method to do something other
-  // than `.append`.
-  attachHtml(collectionView, childView, index) {
-    if (collectionView._isBuffering) {
-      // buffering happens on reset events and initial renders
-      // in order to reduce the number of inserts into the
-      // document, which are expensive.
-      collectionView._bufferedChildren.splice(index, 0, childView);
-    } else {
-      // If we've already rendered the main collection, append
-      // the new child into the correct order if we need to. Otherwise
-      // append to the end.
-      if (!collectionView._insertBefore(childView, index)) {
-        collectionView._insertAfter(childView);
-      }
-    }
-  },
-
-  // Internal method. Check whether we need to insert the view into the correct position.
-  _insertBefore(childView, index) {
-    let currentView;
-    const findPosition = this.sort && (index < this.children.length - 1);
-    if (findPosition) {
-      // Find the view after this one
-      currentView = _.find(this.children._views, (view) => {
-        return view._index === index + 1;
-      });
+      this._destroyChildView(view);
     }
 
-    if (currentView) {
-      this.beforeEl(currentView.el, childView.el);
-      return true;
+    this.stopListening(view);
+  },
+
+  _destroyChildView(view) {
+    if (view._isDestroyed) {
+      return;
     }
 
-    return false;
-  },
-
-  // Override to handle DOM inserting differently
-  beforeEl(el, siblings) {
-    this.$(el).before(siblings);
-  },
-
-  // Internal method. Append a view to the end of the $el
-  _insertAfter(childView) {
-    this.Dom.appendContents(this.el, childView.el, {_$el: this.$el, _$contents: childView.$el});
-  },
-
-  // Internal method to set up the `children` object for storing all of the child views
-  _initChildViewStorage() {
-    this.children = new ChildViewContainer();
+    view._shouldDisableEvents = this.monitorViewEvents === false;
+    destroyView(view);
   },
 
   // called by ViewMixin destroy
   _removeChildren() {
     this._destroyChildren();
+    const emptyRegion = this.getEmptyRegion();
+    emptyRegion.destroy();
+    delete this._addedViews;
   },
 
   // Destroy the child views that this collection view is holding on to, if any
-  _destroyChildren(options) {
-    if (!this.children.length) {
+  _destroyChildren() {
+    if (!this.children || !this.children.length) {
       return;
     }
 
     this.triggerMethod('before:destroy:children', this);
+    if (this.monitorViewEvents === false) {
+      this.Dom.detachContents(this.el, this.$el);
+    }
     _.each(this.children._views, _.bind(this._removeChildView, this));
-    this.children._updateLength();
     this.triggerMethod('destroy:children', this);
-  },
-
-  // Return true if the given child should be shown. Return false otherwise.
-  // The filter will be passed (child, index, collection), where
-  //  'child' is the given model
-  //  'index' is the index of that model in the collection
-  //  'collection' is the collection referenced by this CollectionView
-  _shouldAddChild(child, index) {
-    const filter = this.filter;
-    return !_.isFunction(filter) || filter.call(this, child, index, this.collection);
   }
 }, {
   setDomApi
