+++ test/unit/collection-view/collection-view-data.spec.js
@@ -3,10 +3,10 @@
 import $ from 'jquery';
 import _ from 'underscore';
 import Backbone from 'backbone';
-import CollectionView from '../../../src/next-collection-view';
+import CollectionView from '../../../src/collection-view';
 import View from '../../../src/view';
 
-describe('next CollectionView Data', function() {
+describe('CollectionView Data', function() {
   let MyCollectionView;
 
   beforeEach(function() {
