--- test/unit/sorted-views.spec.js
@@ -1,548 +1,0 @@
-import _ from 'underscore';
-
-describe('collection/composite view sorting', function() {
-  'use strict';
-
-  beforeEach(function() {
-    this.ChildView = Marionette.View.extend({
-      template: _.template('<%= foo %>')
-    });
-
-    this.CollectionView = Marionette.CollectionView.extend({
-      childView: this.ChildView
-    });
-
-    this.CompositeView = Marionette.CompositeView.extend({
-      childView: this.ChildView,
-      template: _.template('<div id="container"></div>')
-    });
-
-    this.collection = new Backbone.Collection([{foo: 1, bar: 4}, {foo: 2, bar: 3}, {foo: 3, bar: 2}]);
-  });
-
-  describe('when working with collections with comparators', function() {
-    beforeEach(function() {
-      this.collection.comparator = 'foo';
-
-      this.collectionView = new this.CollectionView({
-        childView: this.ChildView,
-        collection: this.collection
-      });
-
-      this.compositeView = new this.CompositeView({
-        childView: this.ChildView,
-        collection: this.collection,
-      });
-
-      this.sinon.spy(this.collectionView, 'resortView');
-      this.sinon.spy(this.compositeView, 'resortView');
-
-      this.collectionView.render();
-      this.compositeView.render();
-    });
-
-    describe('when adding a model', function() {
-      beforeEach(function() {
-        this.model = new Backbone.Model({foo: 4, bar: 1});
-        this.collection.add(this.model);
-      });
-
-      it('should add the model to the list', function() {
-        expect(this.collectionView.children).to.have.lengthOf(4);
-      });
-
-      it('should have the order in the dom', function() {
-        expect(this.collectionView.$el).to.have.$text('1234');
-        expect(this.compositeView.$el).to.have.$text('1234');
-      });
-    });
-
-    describe('when adding a model with the "at" option', function() {
-      beforeEach(function() {
-        this.model = new Backbone.Model({foo: 0, bar: 1});
-        this.collection.add(this.model, {at: 5});
-      });
-
-      it('should add the model to the list', function() {
-        expect(this.collectionView.children).to.have.lengthOf(4);
-      });
-
-      it('should ignore the sorted order in the dom', function() {
-        expect(this.collectionView.$el).to.have.$text('1230');
-        expect(this.compositeView.$el).to.have.$text('1230');
-      });
-
-      describe('and adding another', function() {
-        beforeEach(function() {
-          this.model = new Backbone.Model({foo: 5, bar: 0});
-          this.collection.add(this.model);
-        });
-
-        it('should render the sorted order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('01235');
-          expect(this.compositeView.$el).to.have.$text('01235');
-        });
-      });
-    });
-
-    describe('when silently adding a model', function() {
-      beforeEach(function() {
-        this.model = new Backbone.Model({foo: 4, bar: 1});
-        this.collection.add(this.model, {silent: true});
-      });
-
-      it('should not add the model to the list', function() {
-        expect(this.collectionView.children).to.have.lengthOf(3);
-      });
-
-      it('should have the order in the dom', function() {
-        expect(this.collectionView.$el).to.have.$text('123');
-        expect(this.compositeView.$el).to.have.$text('123');
-      });
-
-      describe('and then adding another', function() {
-        beforeEach(function() {
-          this.model = new Backbone.Model({foo: 5, bar: 0});
-          this.collection.add(this.model);
-        });
-
-        it('should add both models to the list', function() {
-          expect(this.collectionView.children).to.have.lengthOf(5);
-        });
-
-        it('should have the order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('12345');
-          expect(this.compositeView.$el).to.have.$text('12345');
-        });
-
-        it('should call resortViews', function() {
-          expect(this.collectionView.resortView).to.have.been.calledOnce;
-          expect(this.compositeView.resortView).to.have.been.calledOnce;
-        });
-
-        describe('and the first, middle, and last models are removed', function() {
-          beforeEach(function() {
-            this.collection.remove([this.collection.at(0), this.collection.at(2), this.collection.at(-1)]);
-          });
-
-          it('should have children view with correct _index', function() {
-            expect(this.collectionView.children.findByModel(this.collection.at(0))).to.have.property('_index', 0);
-            expect(this.collectionView.children.findByModel(this.collection.at(1))).to.have.property('_index', 1);
-          });
-        });
-      });
-    });
-
-    describe('when removing a model', function() {
-      beforeEach(function() {
-        this.model = this.collection.at(1);
-        this.collection.remove(this.model);
-      });
-
-      it('should have the order in the dom', function() {
-        expect(this.collectionView.$el).to.have.$text('13');
-        expect(this.compositeView.$el).to.have.$text('13');
-      });
-
-      it('should have children view with correct _index', function() {
-        expect(this.collectionView.children.findByModel(this.collection.at(0))).to.have.property('_index', 0);
-        expect(this.collectionView.children.findByModel(this.collection.at(1))).to.have.property('_index', 1);
-      });
-
-      describe('and then adding another', function() {
-        beforeEach(function() {
-          this.model = new Backbone.Model({foo: '4'});
-          this.sinon.spy(this.collectionView, 'render');
-          this.collection.add(this.model);
-        });
-
-        it('should have the order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('134');
-          expect(this.compositeView.$el).to.have.$text('134');
-        });
-
-        it('should not call render', function() {
-          expect(this.collectionView.render.callCount).to.equal(0);
-        });
-      });
-    });
-
-    describe('when changing the comparator', function() {
-      beforeEach(function() {
-        this.collection.comparator = 'bar';
-      });
-
-      describe('and triggering a sort', function() {
-        beforeEach(function() {
-          this.collection.sort();
-        });
-
-        it('should have the order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('321');
-          expect(this.compositeView.$el).to.have.$text('321');
-        });
-
-        it('should call resortViews', function() {
-          expect(this.collectionView.resortView).to.have.been.calledOnce;
-          expect(this.compositeView.resortView).to.have.been.calledOnce;
-        });
-      });
-
-      describe('and adding a new child', function() {
-        beforeEach(function() {
-          this.model = new Backbone.Model({foo: '4', bar: '4'});
-          this.collection.add(this.model);
-        });
-
-        it('should have the order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('3214');
-          expect(this.compositeView.$el).to.have.$text('3214');
-        });
-      });
-    });
-  });
-
-  describe('when working with collections with a custom view comparator', function() {
-    beforeEach(function() {
-      this.collection.comparator = 'foo';
-      this.startCollectionOrder = this.collection.map('foo').join();
-
-      this.collectionView = new this.CollectionView({
-        childView: this.ChildView,
-        collection: this.collection,
-        viewComparator: 'bar'
-      });
-
-      this.compositeView = new this.CompositeView({
-        childView: this.ChildView,
-        collection: this.collection,
-        viewComparator: function(modelA, modelB) {
-          return modelA.get('bar') > modelB.get('bar') ? 1 : -1;
-        }
-      });
-
-      this.sinon.spy(this.collectionView, 'resortView');
-      this.sinon.spy(this.compositeView, 'resortView');
-
-      this.collectionView.render();
-      this.compositeView.render();
-      this.endCollectionOrder = this.collection.map('foo').join();
-    });
-
-    it('should not change collection order', function() {
-      expect(this.startCollectionOrder).to.equal(this.endCollectionOrder);
-    });
-
-    describe('when adding a model', function() {
-      beforeEach(function() {
-        this.model = new Backbone.Model({foo: 4, bar: 1});
-        this.collection.add(this.model);
-      });
-
-      it('should add the model to the list', function() {
-        expect(this.collectionView.children).to.have.lengthOf(4);
-      });
-
-      it('should have the order in the dom', function() {
-        expect(this.collectionView.$el).to.have.$text('4321');
-        expect(this.compositeView.$el).to.have.$text('4321');
-      });
-    });
-
-    describe('when silently adding a model', function() {
-      beforeEach(function() {
-        this.model = new Backbone.Model({foo: 4, bar: 1});
-        this.collection.add(this.model, {silent: true});
-      });
-
-      it('should not add the model to the list', function() {
-        expect(this.collectionView.children).to.have.lengthOf(3);
-      });
-
-      it('should have the order in the dom', function() {
-        expect(this.collectionView.$el).to.have.$text('321');
-        expect(this.compositeView.$el).to.have.$text('321');
-      });
-
-      describe('and then adding another', function() {
-        beforeEach(function() {
-          this.model = new Backbone.Model({foo: 5, bar: 0});
-          this.collection.add(this.model);
-        });
-
-        it('should add both models to the list', function() {
-          expect(this.collectionView.children).to.have.lengthOf(5);
-        });
-
-        it('should have the order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('54321');
-          expect(this.compositeView.$el).to.have.$text('54321');
-        });
-
-        it('should call resortViews', function() {
-          expect(this.collectionView.resortView).to.have.been.calledOnce;
-          expect(this.compositeView.resortView).to.have.been.calledOnce;
-        });
-      });
-    });
-
-    describe('when removing a model', function() {
-      beforeEach(function() {
-        this.model = this.collection.at(1);
-        this.collection.remove(this.model);
-      });
-
-      it('should have the order in the dom', function() {
-        expect(this.collectionView.$el).to.have.$text('31');
-        expect(this.compositeView.$el).to.have.$text('31');
-      });
-
-      it('should have children view with correct _index', function() {
-        expect(this.collectionView.children.findByModel(this.collection.at(0))).to.have.property('_index', 1);
-        expect(this.collectionView.children.findByModel(this.collection.at(1))).to.have.property('_index', 0);
-      });
-
-      describe('and then adding another', function() {
-        beforeEach(function() {
-          this.model = new Backbone.Model({foo: '4', bar: '1'});
-          this.sinon.spy(this.collectionView, 'render');
-          this.collection.add(this.model);
-        });
-
-        it('should have the order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('431');
-          expect(this.compositeView.$el).to.have.$text('431');
-        });
-
-        it('should not call render', function() {
-          expect(this.collectionView.render.callCount).to.equal(0);
-        });
-      });
-    });
-
-    describe('when adding a model with the "at" option', function() {
-      beforeEach(function() {
-        this.model = new Backbone.Model({foo: 6, bar: -1});
-        this.collection.add(this.model, {at: 5});
-      });
-
-      it('should add the model to the list', function() {
-        expect(this.collectionView.children).to.have.lengthOf(4);
-      });
-
-      it('should ignore the sorted order in the dom', function() {
-        expect(this.collectionView.$el).to.have.$text('3216');
-        expect(this.compositeView.$el).to.have.$text('3216');
-      });
-
-      describe('and adding another', function() {
-        beforeEach(function() {
-          this.model = new Backbone.Model({foo: 5, bar: 0});
-          this.collection.add(this.model);
-        });
-
-        it('should render the sorted order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('65321');
-          expect(this.compositeView.$el).to.have.$text('65321');
-        });
-      });
-    });
-
-    describe('when changing the comparator', function() {
-      beforeEach(function() {
-        this.collectionView.viewComparator = function(a, b) {
-          if (a.get('foo') < b.get('foo')) { return -1; }
-          if (a.get('foo') > b.get('foo')) { return 1; }
-          return 0;
-        };
-        this.compositeView.viewComparator = 'foo';
-      });
-
-      describe('and triggering a sort', function() {
-        beforeEach(function() {
-          this.collection.sort();
-        });
-
-        it('should have the order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('123');
-          expect(this.compositeView.$el).to.have.$text('123');
-        });
-
-        it('should call resortViews', function() {
-          expect(this.collectionView.resortView).to.have.been.calledOnce;
-          expect(this.compositeView.resortView).to.have.been.calledOnce;
-        });
-      });
-
-      describe('and adding a new child', function() {
-        beforeEach(function() {
-          this.model = new Backbone.Model({foo: '4', bar: '1'});
-          this.collection.add(this.model);
-        });
-
-        it('should have the order in the dom', function() {
-          expect(this.collectionView.$el).to.have.$text('1234');
-          expect(this.compositeView.$el).to.have.$text('1234');
-        });
-      });
-    });
-  });
-
-  describe('when using `{sort: false}`', function() {
-    beforeEach(function() {
-      this.collectionView = new this.CollectionView({
-        childView: this.ChildView,
-        collection: this.collection,
-        sort: false
-      });
-
-      this.compositeView = new this.CompositeView({
-        childView: this.ChildView,
-        collection: this.collection,
-        sort: false
-      });
-
-      this.collectionView.render();
-      this.compositeView.render();
-    });
-
-    describe('and adding a model', function() {
-      beforeEach(function() {
-        this.model = new Backbone.Model({foo: '0', bar: '5'});
-        this.collection.add(this.model);
-      });
-
-      it('should add the model to the list', function() {
-        expect(this.collectionView.children).to.have.lengthOf(4);
-      });
-
-      it('should have the order in the dom', function() {
-        expect(this.collectionView.$el).to.have.$text('1230');
-        expect(this.compositeView.$el).to.have.$text('1230');
-      });
-    });
-  });
-
-  describe('when using `{ reorderOnSort: true }`', function() {
-    var texts = {
-      asOption: 'as an option',
-      onPrototype: 'on the prototype',
-      viewComparator: 'with viewComparator'
-    };
-
-    var getSpecTitle = function(options) {
-      return _.map(options, function(v, k) {
-        return texts[k];
-      }).join(' ');
-    };
-
-    var describeSpec = function(specOptions) {
-      describe(getSpecTitle(specOptions), function() {
-        beforeEach(function() {
-          var commonAttrs = {
-            childView: this.ChildView,
-            collection: this.collection
-          };
-
-          if (specOptions.viewComparator) {
-            commonAttrs.viewComparator = 'foo';
-          }
-
-          if (specOptions.asOption) {
-            this.collectionView = new this.CollectionView(_.extend({}, {
-              reorderOnSort: true
-            }, commonAttrs));
-
-            this.compositeView = new this.CompositeView(_.extend({}, {
-              reorderOnSort: true,
-              childViewContainer: '#container'
-            }, commonAttrs));
-          } else if (specOptions.onPrototype) {
-            var ReorderedCollectionView = this.CollectionView.extend({
-              reorderOnSort: true,
-              onReorder: this.sinon.spy(),
-              onBeforeReorder: this.sinon.spy()
-            });
-
-            var ReorderedCompositeView = this.CompositeView.extend({
-              reorderOnSort: true,
-              childViewContainer: '#container',
-              onReorder: this.sinon.spy(),
-              onBeforeReorder: this.sinon.spy()
-            });
-
-            this.collectionView = new ReorderedCollectionView(commonAttrs);
-            this.compositeView = new ReorderedCompositeView(commonAttrs);
-          }
-
-          this.collectionView.render();
-          this.compositeView.render();
-          this.sinon.spy(this.collectionView, 'reorder');
-          this.sinon.spy(this.collectionView, 'render');
-          this.sinon.spy(this.collectionView, 'trigger');
-
-          var cmp = function(m) {
-            return m.get('bar');
-          };
-
-          if (specOptions.viewComparator) {
-            this.collection.comparator = 'foo';
-            this.collectionView.viewComparator = cmp;
-            this.compositeView.viewComparator = cmp;
-          } else {
-            this.collection.comparator = cmp;
-          }
-          this.collection.sort();
-        });
-
-        it('should call reorder instead of render', function() {
-          expect(this.collectionView.render).not.to.have.been.called;
-          expect(this.collectionView.reorder).to.have.been.calledOnce;
-        });
-
-        it('should reorder the DOM', function() {
-          expect(this.collectionView.$el).to.have.$text('321');
-        });
-
-        it('should triggerMethods events', function() {
-          var cv = this.collectionView;
-          if (specOptions.onPrototype) {
-            expect(cv.onBeforeReorder).calledBefore(cv.onReorder);
-          }
-          expect(cv.trigger).to.have.been.calledWith('before:reorder');
-          expect(cv.trigger).to.have.been.calledWith('reorder');
-          expect(cv.trigger).to.have.been.calledTwice;
-        });
-
-        it('should respect the childViewContainer in a CompositeView', function() {
-          expect(this.compositeView.$('#container')).to.have.$text('321');
-        });
-
-        describe('and reversing the sort', function() {
-          beforeEach(function() {
-            var cmp = function(m) {
-              return m.get('foo');
-            };
-            if (specOptions.viewComparator) {
-              this.collection.comparator = 'bar';
-              this.collectionView.viewComparator = cmp;
-              this.compositeView.viewComparator = cmp;
-            } else {
-              this.collection.comparator = cmp;
-            }
-            this.collection.sort();
-          });
-
-          it('should reorder the DOM', function() {
-            expect(this.collectionView.$el).to.have.$text('123');
-          });
-        });
-      });
-    };
-
-    describeSpec({asOption: true});
-    describeSpec({asOption: true, viewComparator: true});
-    describeSpec({onPrototype: true});
-    describeSpec({onPrototype: true, viewComparator: true});
-  });
-});
